// apps/aptos/utils/swap-get-route/types.ts
export type PoolReserve = {
  type: string
  data: {
    block_timestamp_last: string
    reserve_x: string
    reserve_y: string
  }
}

export type PoolInfo = {
  type: string
  data: {
    decimals: number
    name: string
    symbol: string
    supply: {
      vec: {
        aggregator: { vec: any[] }
        integer: { vec: { limit: string; value: string }[] }
      }[]
    }[]
  }
}

export type Route = {
  route: string[]
  amountOut: number
  priceImpact: number
}

export interface Vertex extends PoolReserve {
  pair: string
  res_x: number
  res_y: number
  lpTokenInfo: PoolInfo
}